Add a response map and dedicated echo function
A response hash map is used to map supported response types to their
responses. This is to ensure that the responses needed can all be
found in the same place, reducing the chances of unintended errors.

An echo function is also added to ensure that all responses we are
trying to give through the response manager are wrapped in the
desired format.