Complete level-2: add functionality to add and list tasks
Mei is now able to receive input to add task and list out all user's
tasks.

Communication between the Task Manager and Response Manager
is established by storing the singleton instances in the respective
classes.

This is so that the Response Manager can process the user's input to
add or list tasks and prompt the Task Manager to execute the related
methods, then prompting the Response Manager to echo the result back
to the user.

There is space for errors and exceptions. Therefore,
Let's
- Add exception handling for task-related methods.
- Brainstorm more scenarios of the user's input to redirect them to
the appropriate methods and make the correct response.