Add functionality to undo the latest command
Caching the latest command by reversing it (e.g. add -> delete,
mark -> unmark and vice versa) will allow execution of this
reversed command if undo is given as the input.

A modification to how the tasks are stored as data by including the
add task command is necessary to make the undo more versatile as we
don't have to manually change how is the add task command defined
for each task types in their classes whenever we decide to
change it.

Let's:
* look into adding functionality to undo multiple commands in a row
instead of just the most recent one, maybe maintain a list instead
of a String.
* possibly refactor some parts of the code and extract them out into
a class.